# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Restaurantes",
        "version": "1.0.0"
    },
    "paths": {
        "/api/restaurantes": {
            "get": {
                "tags": [
                    "Restaurantes"
                ],
                "summary": "Listar restaurantes",
                "operationId": "cb0c09f50d81885b67a3bf8e11a13c2d",
                "responses": {
                    "200": {
                        "description": "Lista de restaurantes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Operacion Exitosa"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "El Cielo"
                                                    },
                                                    "direccion": {
                                                        "type": "string",
                                                        "example": "Calle 10 # 5-67"
                                                    },
                                                    "telefono": {
                                                        "type": "string",
                                                        "example": "3001112233"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-01-01T00:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-01-01T00:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Restaurantes"
                ],
                "summary": "Crear un restaurante",
                "operationId": "e7b8bf55ccd8f5f9756bd048ca133784",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "direccion",
                                    "telefono"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "El Cielo"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Cra 9 # 20-15"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "3001112233"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Restaurante creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Restaurante creado correctamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "El Cielo"
                                                },
                                                "direccion": {
                                                    "type": "string",
                                                    "example": "Calle 10 # 5-67"
                                                },
                                                "telefono": {
                                                    "type": "string",
                                                    "example": "3001112233"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-01T00:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-01T00:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/restaurantes/{id}": {
            "get": {
                "tags": [
                    "Restaurantes"
                ],
                "summary": "Obtener un restaurante por ID",
                "operationId": "38a061ea2dae0380b410123c98a91098",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del restaurante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restaurante encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Operacion Exitosa"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "El Cielo"
                                                },
                                                "direccion": {
                                                    "type": "string",
                                                    "example": "Calle 10 # 5-67"
                                                },
                                                "telefono": {
                                                    "type": "string",
                                                    "example": "3001112233"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-01T00:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-01T00:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Restaurantes"
                ],
                "summary": "Actualizar completamente un restaurante",
                "operationId": "61a07d1d713b0142ddd45b88acca8a28",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "direccion",
                                    "telefono"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "El Cielo"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Cra 9 # 20-15"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "3001112233"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Restaurante actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Restaurante actualizado correctamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "El Cielo"
                                                },
                                                "direccion": {
                                                    "type": "string",
                                                    "example": "Calle 10 # 5-67"
                                                },
                                                "telefono": {
                                                    "type": "string",
                                                    "example": "3001112233"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-01T00:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-01T00:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Restaurantes"
                ],
                "summary": "Eliminar un restaurante",
                "operationId": "b3b0e0a13a84a99b036cafc60ee6c3a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Restaurante eliminado correctamente"
                    },
                    "409": {
                        "description": "Error al eliminar el restaurante"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Restaurantes"
                ],
                "summary": "Actualizar parcialmente un restaurante",
                "operationId": "afd5a6cb4ad086111e0c6b7623220f8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "direccion",
                                    "telefono"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "El Cielo"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Cra 9 # 20-15"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "3001112233"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Restaurante actualizado parcialmente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Restaurante actualizado correctamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "El Cielo"
                                                },
                                                "direccion": {
                                                    "type": "string",
                                                    "example": "Calle 10 # 5-67"
                                                },
                                                "telefono": {
                                                    "type": "string",
                                                    "example": "3001112233"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-01T00:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-01-01T00:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "Clave API para autorización",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Restaurantes",
            "description": "Operaciones sobre restaurantes"
        }
    ],
    "security": [
        {
            "api_key": []
        }
    ]
}